Cairo Port on Mac OSX started and working at the same level than  WIN32 port !
Updated line_style() to support floating point width for subpixel resolution ...

Still needs to refine several aspects:
   - find a way to move cairo context switch from draw() in win32
      because it's not as efficient than synchronize the global cc context only when global gc/hdc changes
   - when testing arc:
       some strokes are missing (see the arcs)
   - default line witdh has subpixel resolution by default which may not be 
       the most adequate default for widgets (too much dithering)
   - use of push/pop cairo context groups api's
     this feature seems not be available on all distrib + it is 
     apparently currently revamped by the cairo team
     we could need this feature to avoid line_width() setting 
     affect the widgets drawing as a side-effect.



git-svn-id: http://seriss.com/public/fltk/fltk/branches/fcupd1@5232 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ cairo/src/win32/setcolor.cxx
@@ -73,7 +73,7 @@ static void free_pen()
   current_pen = 0;
 }
 
-void fltk::line_style(int style, int width, char* dashes) {
+void fltk::line_style(int style, double  width, char* dashes) {
 #if !USE_CAIRO
   static DWORD Cap[4]= {PS_ENDCAP_ROUND, PS_ENDCAP_FLAT, PS_ENDCAP_ROUND, PS_ENDCAP_SQUARE};
   static DWORD Join[4]={PS_JOIN_ROUND, PS_JOIN_MITER, PS_JOIN_ROUND, PS_JOIN_BEVEL};
@@ -89,8 +89,8 @@ void fltk::line_style(int style, int width, char* dashes) {
     lstyle = style & 0xff | Cap[(style>>8)&3] | Join[(style>>12)&3];
   }
   // for some reason zero width does not work at all:
-  if (!width) width = 1;
-  line_width = width;
+  if (!width) width = 1.0;
+  line_width = (int) width;
   if (current_pen) free_pen();
 #else
   int ndashes = dashes ? strlen(dashes) : 0;
@@ -115,12 +115,7 @@ void fltk::line_style(int style, int width, char* dashes) {
       *p++ = dash; *p++ = gap;
       break;
     case DOT:
-      *p++ = dot; *p++ = gap;
-      // Bug in XFree86 3.0? If I only use the above two pieces it does
-      // not completely "erase" the previous dash pattern. Making it longer
-      // like this seems to fix this. For some reason this bug is only for
-      // the dot pattern (not the dash), and only for 0-width lines:
-      *p++ = dot; *p++ = gap; *p++ = dot; *p++ = gap;
+      *p++ = dot; *p++ = gap;  *p++ = dot; *p++ = gap; *p++ = dot; *p++ = gap;
       break;
     case DASHDOT:
       *p++ = dash; *p++ = gap; *p++ = dot; *p++ = gap;
